fix for multiplier overflow for some inputs

+++ rust/src/day3.rs
@@ -29,13 +29,13 @@ fn part_1(map: &Vec<String>) -> u32 {
   count_trees_on_route(map, 3, 1)
 }
 
-fn part_2(map: &Vec<String>) -> u32 {
+fn part_2(map: &Vec<String>) -> u64 {
   let slope1 = count_trees_on_route(map, 1, 1);
   let slope2 = count_trees_on_route(map, 3, 1);
   let slope3 = count_trees_on_route(map, 5, 1);
   let slope4 = count_trees_on_route(map, 7, 1);
   let slope5 = count_trees_on_route(map, 1, 2);
-  slope1 * slope2 * slope3 * slope4 * slope5
+  u64::from(slope1 * slope2) * u64::from(slope3 * slope4 * slope5)
 }
 
 #[cfg(test)]
